% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_chisq.R
\name{plot_chisq}
\alias{plot_chisq}
\title{Plot a Chi-square Distribution with Shading}
\usage{
plot_chisq(df, shadeValues = NULL, col.shade = "cornflowerblue", ...)
}
\arguments{
\item{df}{The degrees of freedom for the chi-square distribution}

\item{shadeValues}{A single value that identifies the boundary of the shaded region}

\item{col.shade}{The color of the shaded region}

\item{...}{Other graphical parameters passed to \link[base]{plot} (e.g.,
\code{xlim}, \code{main}, etc.). Generally, these will not control the
shading, only the normal distribution curve.}
}
\value{
A plot of a chi-square distribution curve with optional shading.
}
\description{
Draw a chi-square distribution curve with appropriate shading to the right
of the test statistic for chi-square hypothesis tests
}
\details{
Since \code{shadeValues} is a
single number, the shaded region will be to the \emph{right} of
\code{shadeValues} since we are interested only in chi-square hypothesis testing.
}
\examples{

# Shade the area for the chi-square test statistic of 4.1 with degrees of freedom 2.
plot_chisq(df = 2, shadeValues = 4.1)

# Shade the area for the chi-square test statistic of 8.223 with degrees of freedom 10.
plot_chisq(df = 10, shadeValues = 8.233)
}
\seealso{
\link{plot_norm}
}
