% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_norm.R
\name{plot_norm}
\alias{plot_norm}
\title{Plot a Normal Distribution with Shading}
\usage{
plot_norm(
  mean = 0,
  sd = 1,
  shadeValues = NULL,
  direction = c("less", "greater", "beyond", "between"),
  col.shade = "cornflowerblue",
  ...
)
}
\arguments{
\item{mean}{The mean of the normal distribution to plot (mu)}

\item{sd}{The standard deviation of the normal distribution to plot (sigma)}

\item{shadeValues}{Either a single number or a vector of two numbers which
identify the boundary or boundaries of the shaded region}

\item{direction}{A character string (in quotes) indicating the direction from
shadeValues to shade under the normal curve. Must be one of "\code{less}",
"\code{greater}", "\code{beyond}", or "\code{between}". See Details below.}

\item{col.shade}{The color of the shaded region}

\item{...}{Other graphical parameters passed to \link[base]{plot} (e.g.,
\code{xlim}, \code{main}, etc.). Generally, these will not control the
shading, only the normal distribution curve.}
}
\value{
A plot of a normal distribution curve with optional shading.
}
\description{
Draw a normal distribution curve and optional shading of a particular region
}
\details{
The \code{direction} argument is used to control the region under
the normal distribution curve which is shaded. If \code{shadeValues} is a
single number, \code{direction} must be either "\code{less}" (in which case
the shaded region will be to the \emph{left} of \code{shadeValues}) or
"\code{greater}" (the shaded region will be to the \emph{right} of
\code{shadeValues}).

If \code{shadeValues} is a \emph{vector}, then \code{direction} must be either "\code{between}"
(the region between the two numbers in \code{shadeValues} will be shaded) or
"\code{beyond}" (the region below the smaller and above the larger of the
\code{shadeValues} will be shaded).
}
\examples{
# Probability of a value being between 3 and 3.5 in a N(3.39, 0.55) distribution
plot_norm(mean = 3.39, sd = 0.55,
          shadeValues = c(3, 3.5),
          direction = "between",
          col.shade = "forestgreen")

# Probability of a value being greater than 1.96 in a N(0, 1) distributio
plot_norm(shadeValues = 1.96,
          direction = "greater",
          col.shade = "peachpuff")
}
