% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_t.R
\name{plot_t}
\alias{plot_t}
\title{Plot a t Distribution with Shading}
\usage{
plot_t(
  df,
  shadeValues = NULL,
  direction = c("less", "greater", "beyond", "between"),
  col.shade = "cornflowerblue",
  ...
)
}
\arguments{
\item{df}{The degrees of freedom of the t distribution to plot}

\item{shadeValues}{Either a single number or a vector of two numbers which
identify the boundary or boundaries of the shaded region}

\item{direction}{A character string (in quotes) indicating the direction from
shadeValues to shade under the normal curve. Must be one of "\code{less}",
"\code{greater}", "\code{beyond}", or "\code{between}". See Details below.}

\item{col.shade}{The color of the shaded region}

\item{...}{Other graphical parameters passed to \link[base]{plot} (e.g.,
\code{xlim}, \code{main}, etc.). Generally, these will not control the
shading, only the normal distribution curve.}
}
\value{
A plot of a normal distribution curve with optional shading.
}
\description{
Draw a t distribution curve and optional shading of a particular region.
}
\details{
The \code{direction} argument is used to control the region under
the normal distribution curve which is shaded. If \code{shadeValues} is a
single number, \code{direction} must be either "\code{less}" (in which case
the shaded region will be to the \emph{left} of \code{shadeValues}) or
"\code{greater}" (the shaded region will be to the \emph{right} of
\code{shadeValues}).

If \code{shadeValues} is a \emph{vector}, then \code{direction} must be either "\code{between}"
(the region between the two numbers in \code{shadeValues} will be shaded) or
"\code{beyond}" (the region below the smaller and above the larger of the
\code{shadeValues} will be shaded).
}
\examples{
# Shade the region below 1.4 on the t(4) distribution.
plot_t(df = 4, shadeValues = 1.4, direction = "less")

# Shade the region between -2 and 0.5 on the t(13) distribution.
plot_t(df = 13, shadeValues = c(-2, 0.5), direction = "between")
}
\seealso{
\link{plot_norm}
}
