% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffle_two_groups.R
\name{shuffle_two_groups}
\alias{shuffle_two_groups}
\title{Randomization for Two-Group Comparisons}
\usage{
shuffle_two_groups(d, groups)
}
\arguments{
\item{d}{A data frame that contains the}

\item{groups}{The name of the variable (in quotes!) that identifies which
group each observation belongs to}
}
\value{
The original data frame with the group labels randomly shuffled
among observations
}
\description{
Shuffles data for use in two-group simulations.
}
\examples{
# Make a data frame to shuffle
og_data <- data.frame(
  "group" = c(rep("Group A", 600), rep("Group B", 200)),
  "result" = c(rep("Failure", 402), rep("Success", 198),
               rep("Failure", 120), rep("Success", 80)),
  stringsAsFactors = TRUE)

# Tabulate successes & failures in each group in original data
table(og_data$result, og_data$group)

# Shuffle data with shuffle_two_groups
shuffle_data <- shuffle_two_groups(og_data, "group")

# Tabulate successes & failures in each group in shuffled data
table(shuffle_data$result, shuffle_data$group)

}
