% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_chance_model.R
\name{simulate_chance_model}
\alias{simulate_chance_model}
\title{Simulate from a Chance Model}
\usage{
simulate_chance_model(chanceSuccess, numDraws, numRepetitions)
}
\arguments{
\item{chanceSuccess}{A number between 0 and 1 representing the probability of
observing a "success" (however defined). Can be a decimal or a fraction;
e.g., 0.55 or 55/100.}

\item{numDraws}{The number of times to draw a poker chip from the bag needed
to complete one repetition of the simulation. Must be a positive, whole
number; if not, it will be rounded to the nearest whole number.}

\item{numRepetitions}{The number of times to repeat the simulation process.
Must be a positive, whole number; if not, it will be rounded to the nearest
whole number.}
}
\value{
A vector of simulated proportions of successes. The vector will be
the same length as \code{numRepetitions}.
}
\description{
Simulate from a Chance Model
}
\examples{
# Run 100 repetitions of a simulation in which 26 fair coins are flipped,
# where "success" is heads. Store the results as `x`.
x <- simulate_chance_model(chanceSuccess = 0.5, numDraws = 26, numRepetitions = 100)

# Make a histogram of the results
hist(x)
}
